moved [un]subbed_keyword into helpers.py
